python_bot/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.1.0
audioop-lts==0.2.1
certifi==2025.1.31
cffi==1.17.1
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
outcome==1.3.0.post0
propcache==0.2.1
pycparser==2.22
PyNaCl==1.5.0
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
yt-dlp==2025.1.26
